requirements.txt
allure-pytest
Faker
pytest
python-dateutil
python-dotenv
requests
selenium
webdriver-manager
